Refactor route structure
